# Fully-Sparse wavenumber arrays