--- a/radis/lbl/broadening.py
+++ b/radis/lbl/broadening.py
@@ -1941,7 +1941,9 @@ class BroadenFactory(BaseFactory):
                 ::
                     DLM_ranges_00 = get_non_zero_wranges(groupby_parameters=["li0", "mi0"], max_range=len(wavenumber_calc))
                 """
-                dgb = df.groupby(groupby_parameters)
+                # EP 31/10: the for loop over df.groupby is the current bottleneck
+                # (not even the sparse_add_at function !)
+                dgb = df.groupby(groupby_parameters, sort=False)
                 DLM_ranges = {}
                 DLM_reduced = {}
                 for groupby_param, group in dgb: