<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

### Todo

- [x] implement fast merging of ranges (avoid computing all booleans arrays 4x times per element of the DIT grid) --> DONE, but slower
- [x] if not enough to improve performances, finish benchmarks to see when to switch on/off the sparse implementation --> DONE, with "auto" mode 
- [ ] keep the reduce/expand optimization of #386 even in non-sparse mode ?
 
### Description
<!-- Provide a general description of what your pull request does. -->

Following a partial sparse implementation of DLM in #386 , which used a reduce/expand step of the wavenumber arrays before/after the convolution step; here I introduce a fully-sparse architecture for the DLM arrays. 
Instead of storing wavenumber arrays, we only store coordinates of the future non-zero ranges.

- The hard part is that the mapping of the lines on the grid (former add-at step) has to be done at the same time as the construction of the sparse ranges. 
- An additional difficulty is that each line contributes to 2x2 sparse-ranges in the LDM grid, which are likely different from one another. Here we use boolean arrays and combine them.

The broadening step of the DLM can be summarized as : 

1. (same as before) find indices of all lines in the (w, wG, wL) grid 
2. (new) grow the wavenumber grids as we add the lines on it. This uses a Pandas DataFrame to first [group lines by lineshape](https://github.com/radis/radis/pull/388/files#diff-85a035414f3d48ffc0db8079be5cac5553de6dd376e4844da9917ffa289dfb64R1949) ( It is the current performance bottleneck, and may be replaced by a custom groupy function. ),  and then a [sparse_add_at](https://github.com/radis/radis/pull/388/files#diff-fd7b5829ec3e91a3bb58217dbf7acde5e9e72391a1b47718b01431faaa9bbe2dR699) function  (which works quite efficiently)
3. (same as #386) convolve only the reduced array


PS : as already mentioned on the Slack, this implementation allows to have a different truncation for each point of the DLM grid and therefore makes it very possible to deploy multiple-wavenumber-grids (soon !)


### NO Test

So far, the implementation is slightly slower than #386. The Distribution (step 2 above) is quite slow, because of Panda's Groupby(). 

```python

import astropy.units as u
from radis import SpectrumFactory, config, plot_diff

print("Without Sparse Range Optimization")
print("=================================")
print("\n"*2)

truncation = 1.5  # cm-1

config["SPARSE_WAVERANGE"] = False
sf = SpectrumFactory(1, 3000,  # 26777,         # cm-1
                  molecule='NO',
                  isotope='1,2,3',
                  truncation=truncation,      #   ~
                  pressure=760 * u.torr,
                  verbose=False,
                  wstep=0.01,
                  mole_fraction=1)

sf.fetch_databank('hitemp')
s = sf.eq_spectrum(700)
s.print_perf_profile()


print("With Sparse Range Optimization")
print("==============================")
print("\n"*2)

config["SPARSE_WAVERANGE"] = True
s2 = sf.eq_spectrum(700)
s2.print_perf_profile()

assert 2*sf.df1.hwhm_voigt.max()*10 < truncation  # HAPI conditions

#%%
s.name = f'0.10.3 ({s.conditions["calculation_time"]:.0f}s)'
s2.name = f'Sparse optim ({s2.conditions["calculation_time"]:.0f}s)'
_, [ax0, ax1] = plot_diff(s, s2)
ax0.set_yscale('log')
```


Output:
```
Without Sparse Range Optimization
=================================


NOspectrum3088769491008 profiler :
    spectrum_calculation      13.645s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.011s 
            copy_database                   0.008s 
            memory_usage_warning            0.003s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.032s 
        calc_lineshift                  0.002s 
        calc_hwhm                       0.058s 
        generate_wavenumber_arrays      0.002s 
        calc_line_broadening            13.496s ███████████████
            precompute_DLM_lineshapes       0.058s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.037s 
            DLM_Distribute_lines            0.513s 
            DLM_convolve                    12.754s ██████████████
            others                          0.134s 
        calc_other_spectral_quan        0.040s 
        generate_spectrum_obj           0.004s 
        others                          0.135s 
With Sparse Range Optimization
==============================



NOspectrum3088769491392 profiler :
    spectrum_calculation      9.275s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.013s 
            copy_database                   0.012s 
            memory_usage_warning            0.001s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.032s 
        calc_lineshift                  0.003s 
        calc_hwhm                       0.060s 
        generate_wavenumber_arrays      0.002s 
        calc_line_broadening            9.125s ███████████████
            precompute_DLM_lineshapes       0.061s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.043s 
            DLM_Distribute_lines            2.758s ████                  # ⬅⬅ this should be improved 
            DLM_convolve                    6.177s ██████████     # ⬅ was ~3.4  in #386 only ? 
            others                          0.086s 
        calc_other_spectral_quan        0.034s 
        generate_spectrum_obj           0.007s 
        others                          0.086s 

```

Match is excellent : 

![image](https://user-images.githubusercontent.com/16088743/139602832-fb2c3d0e-0501-48e0-9420-17980c4aab86.png)

### CO2-Full-Range Test


So performances are not as good as in #386 yet (_EDIT : Fixed, see comments below_) , however; the Memory efficiency is much better as the wavenumbers are stored sparse. This allowed me to compute full-range NO spectra and full-range CO2 spectra ( 9.9 M lines, 2.7M spectral-points )  on a 16 GB RAM laptop without any problem .
Ex : 

```python

import astropy.units as u
from radis import calc_spectrum
s = calc_spectrum(1, 26777,         # cm-1           # <<< full spectral range!
                  molecule='CO2',
                  isotope='1,2,3',
                  pressure=760 * u.torr,
                  Tgas=700,           # K
                  mole_fraction=1,
                  wstep=0.01,
                  databank='hitemp',
                  verbose=3,
                  cutoff=0,
                  
                  )
s.print_perf_profile()
s.name = f"CO2 ({s.conditions['lines_calculated']:.0e}M lines, {s.conditions['calculation_time']:.1f}s)"
s.plot(yscale='log')
import matplotlib.pyplot as plt
plt.legend()
```

```
CO2spectrum3089228638272 profiler :
    spectrum_calculation      101.857s ████████████████
        check_line_databank             0.014s 
        reinitialize                    0.014s 
            copy_database                   0.000s 
            memory_usage_warning            0.013s 
            reset_population                0.001s 
        scaled_eq_linestrength          0.955s 
        calc_lineshift                  0.072s 
        calc_hwhm                       1.580s 
        generate_wavenumber_arrays      0.027s 
        calc_line_broadening            98.719s ███████████████
            precompute_DLM_lineshapes       0.820s 
            DLM_Initialized_vectors         0.001s 
            DLM_closest_matching_line       5.340s 
            DLM_Distribute_lines            38.251s ██████
            DLM_convolve                    53.500s ████████
            others                          0.807s 
        calc_other_spectral_quan        0.406s 
        generate_spectrum_obj           0.041s 
        others                          0.837s 
```
![image](https://user-images.githubusercontent.com/16088743/139603142-99625ee9-4306-4274-b47d-f2aaeee2736d.png)

(So ~ 2.5e11 points*lines / s ; which is what we achieved in the DIT paper ... but we could never scale to these large spectral ranges until now )