@erwanp 2021-10-31
Implementation was right, I had just made a small typo while declaring boolean. 
Fixed --> Extreme performance improvement on the broadening step


### NO 
```
NOspectrum1964030701632 profiler :
    spectrum_calculation      2.868s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.011s 
            copy_database                   0.010s 
            memory_usage_warning            0.001s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.025s 
        calc_lineshift                  0.003s 
        calc_hwhm                       0.058s 
        generate_wavenumber_arrays      0.002s 
        calc_line_broadening            2.735s ███████████████
            precompute_DLM_lineshapes       0.065s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.038s 
            DLM_Distribute_lines            2.213s ████████████
            DLM_convolve                    0.402s ██
            others                          0.017s 
        calc_other_spectral_quan        0.031s 
        generate_spectrum_obj           0.004s 
        others                          0.017s 
```

![image](https://user-images.githubusercontent.com/16088743/139603857-93cff092-2510-4df0-98d4-8832af4ea169.png)



### CO2 

```
CO2spectrum1965157055296 profiler :
    spectrum_calculation      30.009s ████████████████
        check_line_databank             0.042s 
        reinitialize                    0.049s 
            copy_database                   0.001s 
            memory_usage_warning            0.041s 
            reset_population                0.005s 
            others                          0.002s 
        scaled_eq_linestrength          0.998s 
        calc_lineshift                  0.070s 
        calc_hwhm                       1.597s 
        generate_wavenumber_arrays      0.023s 
        calc_line_broadening            26.775s ██████████████
            precompute_DLM_lineshapes       1.020s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       4.911s ██
            DLM_Distribute_lines            18.099s █████████
            DLM_convolve                    2.460s █
            others                          0.285s 
        calc_other_spectral_quan        0.352s 
        generate_spectrum_obj           0.042s 
        others                          0.348s 
```

![image](https://user-images.githubusercontent.com/16088743/139603845-eafbeca3-c7f7-44dc-a755-1200def331e3.png)