@erwanp 2021-11-01
Only warning : unlike #386 where the overhead was reduced to <10%, here there is still some significant overhead (+100%) for dense spectra. See CO2 example : 

```python

import astropy.units as u
from radis import SpectrumFactory, config, plot_diff

print("Without Sparse Range Optimization")
print("=================================")
print("\n"*2)

truncation = 1.5  # cm-1

config["SPARSE_WAVERANGE"] = False
sf = SpectrumFactory(4000 * u.nm, 5000 * u.nm,  # 26777,         # cm-1
                  molecule='CO2',
                  isotope='1,2,3',
                  truncation=truncation,      #   ~
                  pressure=760 * u.torr,
                  verbose=False,
                  wstep=0.01,
                  mole_fraction=1,
                  cutoff=0)

sf.fetch_databank('hitemp')
s = sf.eq_spectrum(700)
s.print_perf_profile()


print("With Sparse Range Optimization")
print("==============================")
print("\n"*2)

config["SPARSE_WAVERANGE"] = True
s2 = sf.eq_spectrum(700)
s2.print_perf_profile()

assert 2*sf.df1.hwhm_voigt.max()*10 < truncation  # HAPI conditions

#%%
s.name = f'0.10.3 ({s.conditions["calculation_time"]:.1f}s)'
s2.name = f'Sparse optim ({s2.conditions["calculation_time"]:.1f}s)'
_, [ax0, ax1] = plot_diff(s, s2, wunit='nm')
ax0.set_yscale('log')

```
![image](https://user-images.githubusercontent.com/16088743/139610000-3b3dc114-562e-4cd9-b877-36f15914ca25.png)
  
Which is due to the large DLM Distribution Step
```

CO2spectrum2829470404800 profiler :
    spectrum_calculation      1.063s ████████████████
        check_line_databank             0.004s 
        reinitialize                    0.089s █
            copy_database                   0.084s █
            memory_usage_warning            0.005s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.165s ██
        calc_lineshift                  0.020s 
        calc_hwhm                       0.340s █████
        generate_wavenumber_arrays      0.000s 
        calc_line_broadening            0.440s ██████
            precompute_DLM_lineshapes       0.015s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.262s ███
            DLM_Distribute_lines            0.103s █
            DLM_convolve                    0.045s 
            others                          0.015s 
        calc_other_spectral_quan        0.005s 
        generate_spectrum_obj           0.000s 
        others                          0.015s 
With Sparse Range Optimization
==============================



CO2spectrum2829470405056 profiler :
    spectrum_calculation      2.327s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.120s 
            copy_database                   0.118s 
            memory_usage_warning            0.002s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.178s █
        calc_lineshift                  0.016s 
        calc_hwhm                       0.322s ██
        generate_wavenumber_arrays      0.002s 
        calc_line_broadening            1.684s ███████████
            precompute_DLM_lineshapes       0.016s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.243s █
            DLM_Distribute_lines            1.339s █████████
            DLM_convolve                    0.052s 
            others                          0.035s 
        calc_other_spectral_quan        0.005s 
        generate_spectrum_obj           0.000s 
        others                          0.035s 
```