@dcmvdbekerom 2021-11-01
I copied your NO-test and get similar reductions, but Spyder crashes once it tries to show the figure..
Note also that it complains again about group 6 and class 3 not being implemented
```
Without Sparse Range Optimization
=================================



parse_local_quanta not implemented for molecules of HITRAN group 6
parse_global_quanta not implemented for molecules of HITRAN class 3
c:\users\dcmvd\documents\github\radis\radis\misc\warning.py:350: AccuracyWarning: Some lines are too narrow (FWHM ~ 0.015 cm⁻¹) for the current spectral grid (wstep=0.01). Please reduce wstep to below 0.0051 cm⁻¹. You can use wstep='auto' to get the optimal spectral grid value. You can also ignore by setting `warnings={'AccuracyWarning':'ignore'}` or change the 'GRIDPOINTS_PER_LINEWIDTH_WARN_THRESHOLD' key of radis.config / your ~/radis.json (if you know what you're doing!)
  warnings.warn(WarningType(message))
NOspectrum2472861511360 profiler :
    spectrum_calculation      9.197s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.006s 
            copy_database                   0.005s 
            memory_usage_warning            0.001s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.020s 
        calc_lineshift                  0.002s 
        calc_hwhm                       0.046s 
        generate_wavenumber_arrays      0.003s 
        calc_line_broadening            9.091s ███████████████
            precompute_DLM_lineshapes       0.046s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.030s 
            DLM_Distribute_lines            0.293s 
            DLM_convolve                    8.598s ██████████████
            others                          0.124s 
        calc_other_spectral_quan        0.025s 
        generate_spectrum_obj           0.004s 
        others                          0.125s 
With Sparse Range Optimization
==============================



c:\users\dcmvd\documents\github\radis\radis\misc\warning.py:350: AccuracyWarning: Some lines are too narrow (FWHM ~ 0.015 cm⁻¹) for the current spectral grid (wstep=0.01). Please reduce wstep to below 0.0051 cm⁻¹. You can use wstep='auto' to get the optimal spectral grid value. You can also ignore by setting `warnings={'AccuracyWarning':'ignore'}` or change the 'GRIDPOINTS_PER_LINEWIDTH_WARN_THRESHOLD' key of radis.config / your ~/radis.json (if you know what you're doing!)
  warnings.warn(WarningType(message))
NOspectrum2472988737856 profiler :
    spectrum_calculation      1.712s ████████████████
        check_line_databank             0.000s 
        reinitialize                    0.009s 
            copy_database                   0.007s 
            memory_usage_warning            0.001s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.019s 
        calc_lineshift                  0.002s 
        calc_hwhm                       0.040s 
        generate_wavenumber_arrays      0.001s 
        calc_line_broadening            1.613s ███████████████
            precompute_DLM_lineshapes       0.041s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.029s 
            DLM_Distribute_lines            1.299s ████████████
            DLM_convolve                    0.231s ██
            others                          0.013s 
        calc_other_spectral_quan        0.026s 
        generate_spectrum_obj           0.003s 
        others                          0.013s 
```